Separated models to another namespace
